+++ lib/active_merchant/billing/first_data.rb
@@ -1,0 +1,9 @@
+require "active_merchant/billing/first_data/gateway"
+require "active_merchant/billing/first_data/response"
+
+module ActiveMerchant #:nodoc:
+  module Billing #:nodoc:
+    module FirstData
+    end
+  end
+end
