+++ lib/active_merchant/billing/first_data/response.rb
@@ -1,0 +1,8 @@
+module ActiveMerchant #:nodoc:
+  module Billing #:nodoc:
+    module FirstData
+      class Response
+      end
+    end
+  end
+end
