+++ lib/active_merchant_first_data.rb
@@ -1,1 +1,1 @@
-require "active_merchant/billing/first_data_gateway"
+require "active_merchant/billing/first_data"
